ci: bump android cross version and build in releases (#1914)
Realized I never actually build it in my build CI. This also updates the git version of cross used. I may end up using this for everything to be honest, since the stable version of cross used is ancient.

Note that this is built like it is in normal CI, and that is with no default features.

This PR also enables loong in normal CI, which I forgot to do earlier.